Move ElementContent to src/types
