+++ src/helpers/types/index.ts
@@ -1,3 +1,2 @@
 export * from './VibeComponent';
 export * from './VibeComponentProps';
-export * from './ElementContent';
